Provide Support for Fennel
This commit does not alter the functionality of lovebird if fennel is not present.

Introduced a new function lovebird.loadstringAlt. By default it points to the same function as lovebird.loadstring, but if fennel is not nil it points to  a new function described below. loadstringAlt is only used in the "index" page.

loadstringAlt

If fennel is not nil in the global space, opt to use fennel.compile-string and loadstring in place of raw loadstring

If fennel is not nil, always print the result using fennel.view.